+++ src/app/(header-footer)/signin/components/SocialLogin.tsx
@@ -1,5 +1,5 @@
-import { googleLogin, kakaoLogin } from '@/app/signin/action';
-import { SocialLoginFormBtn } from '@/app/signin/components/SocialLoginFormBtn';
+import { googleLogin, kakaoLogin } from '@/app/(header-footer)/signin/action';
+import { SocialLoginFormBtn } from '@/app/(header-footer)/signin/components/SocialLoginFormBtn';
 import { MESSAGES, SOCIAL_LOGIN_BUTTON } from '@/constants/login';
 
 export function SocialLogin() {
